refactor: change attributes data types from double to float

+++ ConcreteDevices/Lighting.java
@@ -3,14 +3,14 @@ package ConcreteDevices;
 import Devices.BaseDevice;
 
 public class Lighting extends BaseDevice{
-    private double brightness;
+    private float brightness;
     private final String BRIGHTNESS_VALUE_ERROR = "Brightness value should be between 0 and 100";
 
-    public double getBrightness() {
+    public float getBrightness() {
         return this.brightness;
     }
 
-    public void setBrightness(double brightness) {
+    public void setBrightness(float brightness) {
         if (brightness >= 0 && brightness <= 100) {
             this.brightness = brightness;
         }else {
